Recombined files for generating csv.
